ref(quotas): Move quota config to base module
