+++ src/sentry/quotas/redis.py
@@ -5,114 +5,14 @@ import six
 
 from time import time
 
-from sentry.quotas.base import NotRateLimited, Quota, RateLimited
+from sentry.quotas.base import NotRateLimited, Quota, QuotaConfig, RateLimited
 from sentry.utils.redis import (
     get_dynamic_cluster_from_options,
     validate_dynamic_cluster,
     load_script,
 )
-from sentry.utils.json import prune_empty_keys
 
 is_rate_limited = load_script("quotas/is_rate_limited.lua")
-
-
-class BasicRedisQuota(object):
-    """
-    A quota in the most abstract sense consists of an identifier (such as
-    `"organization quota"`, `"smart limit"`, etc) and two integers `limit` and
-    `window` (we accept "limit" events per "window" seconds).
-
-    Sentry applies multiple quotas to an event before accepting it, some of
-    which can be configured by the user depending on plan. An event will be
-    counted against all quotas. For example:
-
-    * If Sentry is told to apply two quotas "one event per minute" and "9999999
-      events per hour", it will practically accept only one event per minute
-    * If Sentry is told to apply "one event per minute" and "30 events per
-      hour", we will be able to get one event accepted every minute. However, if
-      we do that for 30 minutes (ingesting 30 events), we will not be able to get
-      an event through for the rest of the hour. (This example assumes that we
-      start sending events exactly at the start of the time window)
-
-    A `BasicRedisQuota` is a specific quota type that also includes some
-    attributes necessary for looking up event counters and refunds in Redis.
-
-    The `BasicRedisQuota` object is not persisted in any way. It is just a
-    function argument passed around in code. Most importantly, a
-    `BasicRedisQuota` instance does not contain information about how many
-    events can still be accepted (that stuff is stored in Redis but isn't typed
-    out), it only represents settings that should be applied.
-    """
-
-    __slots__ = ["prefix", "subscope", "limit", "window", "reason_code"]
-
-    def __init__(self, prefix=None, subscope=None, limit=None, window=None, reason_code=None):
-        if limit == 0:
-            assert prefix is None and subscope is None, "zero-sized quotas are not tracked in redis"
-            assert window is None, "zero-sized quotas cannot have a window"
-        else:
-            assert prefix, "measured quotas need a prefix to run in redis"
-            assert window and window > 0, "window cannot be zero"
-
-        self.prefix = prefix
-        self.subscope = subscope
-        # maximum number of events in the given window
-        #
-        # None indicates "unlimited amount"
-        # 0 indicates "reject all"
-        # NOTE: Use `quotas.base._limit_from_settings` to map from settings
-        self.limit = limit
-        # time in seconds that this quota reflects
-        self.window = window
-        # a machine readable string
-        self.reason_code = reason_code
-
-    @classmethod
-    def reject_all(cls, reason_code):
-        """
-        A zero-sized quota, which is never counted in Redis. Unconditionally
-        reject the event.
-        """
-
-        return cls(limit=0, reason_code=reason_code)
-
-    @classmethod
-    def limited(cls, prefix, limit, window, reason_code, subscope=None):
-        """
-        A regular quota with limit.
-        """
-
-        assert limit and limit > 0
-        return cls(
-            prefix=prefix, limit=limit, window=window, reason_code=reason_code, subscope=subscope
-        )
-
-    @classmethod
-    def unlimited(cls, prefix, window, subscope=None):
-        """
-        Unlimited quota that is still being counted.
-        """
-
-        return cls(prefix=prefix, window=window, subscope=subscope)
-
-    @property
-    def should_track(self):
-        """
-        Whether the quotas service should track this quota at all.
-        """
-
-        return self.prefix is not None
-
-    def to_json(self):
-        return prune_empty_keys(
-            {
-                "prefix": six.text_type(self.prefix) if self.prefix is not None else None,
-                "subscope": six.text_type(self.subscope) if self.subscope is not None else None,
-                "limit": self.limit,
-                "window": self.window,
-                "reasonCode": self.reason_code,
-            }
-        )
 
 
 class RedisQuota(Quota):
@@ -165,7 +65,7 @@ class RedisQuota(Quota):
         pquota = self.get_project_quota(project)
         if pquota[0] is not None:
             results.append(
-                BasicRedisQuota.limited(
+                QuotaConfig.limited(
                     prefix="p",
                     subscope=project.id,
                     limit=pquota[0],
@@ -177,7 +77,7 @@ class RedisQuota(Quota):
         oquota = self.get_organization_quota(project.organization)
         if oquota[0] is not None:
             results.append(
-                BasicRedisQuota.limited(
+                QuotaConfig.limited(
                     prefix="o", limit=oquota[0], window=oquota[1], reason_code="org_quota"
                 )
             )
@@ -186,7 +86,7 @@ class RedisQuota(Quota):
             kquota = self.get_key_quota(key)
             if kquota[0] is not None:
                 results.append(
-                    BasicRedisQuota.limited(
+                    QuotaConfig.limited(
                         prefix="k",
                         subscope=key.id,
                         limit=kquota[0],
